+++ cauchy-consensus/Cargo.toml
@@ -8,7 +8,7 @@ edition = "2018"
 rayon = "1.3.0"
 
 [dev-dependencies]
-criterion = "0.3.1"
+criterion = "0.3.2"
 rand = "0.7.3"
 
 [[bench]]
