+++ cauchy-crypto/Cargo.toml
@@ -5,4 +5,4 @@ authors = ["Harry Barber <harrybarber@protonmail.com>"]
 edition = "2018"
 
 [dependencies]
-blake3 = "0.2.0"
+blake3 = "0.3.3"
