+++ cauchy-miner/Cargo.toml
@@ -7,9 +7,9 @@ edition = "2018"
 [dependencies]
 crypto = { package = "cauchy-crypto", path = "../cauchy-crypto" }
 futures-channel = "0.3.4"
-parking_lot = "0.10.0"
+parking_lot = "0.10.2"
 rayon = "1.3.0"
 
 [dev-dependencies]
-tokio = { version = "0.2.11", features = ["macros", "rt-core"] }
+tokio = { version = "0.2.20", features = ["macros", "rt-core"] }
 futures-util = "0.3.4"
