+++ cauchy-services/Cargo.toml
@@ -9,9 +9,9 @@ consensus = { package = 'cauchy-consensus',  path = '../cauchy-consensus' }
 crypto = { package = 'cauchy-crypto',  path = '../cauchy-crypto' }
 miner = { package = 'cauchy-miner',  path = '../cauchy-miner' }
 network = { package = 'cauchy-network',  path = '../cauchy-network' }
-rpc = { package = 'cauchy-rpc',  path = '../cauchy-rpc' }
 
 bytes = "0.5.4"
+dashmap = "3.11.1"
 evmap = "9.0.1"
 futures = "0.3.4"
 rand = "0.7.3"
