+++ cauchy-services/src/lib.rs
@@ -1,3 +1,4 @@
 pub mod arena;
 pub mod database;
 pub mod peers;
+pub mod player;
