+++ cauchy-vm/Cargo.toml
@@ -7,12 +7,12 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-bytes = "*"
+bytes = "0.5.4"
 rust-wasm = {git="https://github.com/TheRealBluesun/rust-wasm", tag="v0.1.2"}
-digest = "*"
-rayon = "*"
+digest = "0.8.1"
+rayon = "1.3.0"
 
 [dev-dependencies]
-hex = "*"
-sha2 = "*"
-blake3 = "*"
+hex = "0.4.2"
+sha2 = "0.8.1"
+blake3 = "0.3.3"
