+++ cauchy/Cargo.toml
@@ -17,6 +17,6 @@ clap = { version = "2.33.0", features = ["yaml"] }
 config = "0.10.1"
 dirs = "2.0.2"
 futures = "0.3.4"
-serde = { version = "1.0.104", features = ['derive'] }
+serde = { version = "1.0.106", features = ['derive'] }
 stream-cancel = "0.5.2"
-tokio = { version = "0.2.11", features = ['macros', 'sync'] }
+tokio = { version = "0.2.20", features = ['macros', 'sync'] }
